+++ benchmarks/internal/cluster-worker.mjs
@@ -3,13 +3,12 @@ import { ClusterWorker } from '../../lib/index.mjs'
 import { executeTaskFunction } from '../benchmarks-utils.mjs'
 import { TaskFunctions } from '../benchmarks-types.mjs'
 
-const debug = false
-
 const taskFunction = data => {
   data = data || {}
   data.function = data.function || TaskFunctions.jsonIntegerSerialization
+  data.debug = data.debug || false
   const res = executeTaskFunction(data)
-  debug === true && console.debug(`This is the main thread ${isPrimary}`)
+  data.debug === true && console.debug(`This is the main thread ${isPrimary}`)
   return res
 }